Moved the file mapping logic from the fragment to the view model. Fixed the last row not showing because the data size value didn't reset when changing folders.
